Add files via upload
This a program that focus more on if and else statements.
It is generally a program that helps us to guess a number which
it has input from the user before. There are different cases where the
guessed and the input number can fail. So give this a try for sure.